@@ -16,12 +16,13 @@
         }
     ],
     "require": {
-        "php": "^7.1",
-        "illuminate/support": "^6.0|^7.0|^8.0",
-        "livewire/livewire": "^2.0"
+        "php": "^7.4|^8.0|^8.1",
+        "illuminate/support": "^6.0|^7.0|^8.0|^9.0",
+        "livewire/livewire": "^2.0",
+        "ext-json": "*"
     },
     "require-dev": {
-        "orchestra/testbench": "^5.0|^6.0",
+        "orchestra/testbench": "^5.0|^6.0|^7.0",
         "phpunit/phpunit": "^9.0"
     },
     "autoload": {
